requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
aspose-cells==22.10.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
autograd==1.7.0
autograd-gamma==0.5.0
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
branca==0.6.0
buffer==0.2.1
cachetools==5.3.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==2.0.12
chartjs==1.2
charts==0.4.6
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
colorama==0.4.4
cryptography==41.0.7
cycler==0.11.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==2.0.7
django-shortcuts==1.6
dm-tree==0.1.8
docstring_parser==0.16
entrypoints==0.4
env==0.1.0
et-xmlfile==1.1.0
etils==1.10.0
excel2json-3==0.1.6
executing==0.8.3
Faker==33.1.0
fastjsonschema==2.16.2
filelock==3.8.2
Fiona==1.9.4.post1
Flask==2.2.2
flatbuffers==24.3.25
folium==0.12.1
fonttools==4.33.3
formulaic==1.1.1
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.5.4
geographiclib==2.0
geopandas==0.13.2
geopy==2.3.0
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.25.2
idna==3.3
imbalanced-learn==0.13.0
immutabledict==4.2.0
importlib-metadata==6.8.0
importlib_resources==6.4.5
imutils==0.5.4
interface-meta==1.3.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
json5==0.9.10
jsonschema==4.17.0
jupyter-client==7.3.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter-server==1.21.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.2
jupyterthemes==0.20.0
keras==3.6.0
keras-nightly==3.7.0.dev2024111103
kiwisolver==1.4.2
lesscpy==0.15.1
libclang==18.1.1
lifelines==0.30.0
lightgbm==4.6.0
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
matplotlip==0.2
mdurl==0.1.2
mistune==2.0.4
ml_dtypes==0.5.0
multidict==6.1.0
mysql-connector-python==8.3.0
namex==0.0.8
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.5
nltk==3.9.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
opencage==3.0.2
opencv-python==4.8.0.76
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2022.12.19
platformdirs==2.6.0
plotly==5.8.0
plotly-express==0.4.1
ply==3.11
point==0.0.1
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.29
propcache==0.2.0
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==8.2.2
pygame==2.5.2
Pygments==2.17.2
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.0.9
PyPDF2==3.0.1
pypng==0.20220715.0
pyppeteer==1.0.2
pyproj==3.6.0
PyQRCode==1.2.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==1.0.2
python-telegram-bot==12.8
pytube==12.1.3
pytz==2022.1
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.4.2
regex==2024.11.6
render==1.0.0
reportlab==4.0.7
requests==2.32.3
rich==13.7.0
scikit-learn==1.3.2
scipy==1.11.2
seaborn==0.12.1
Send2Trash==1.8.0
shapely==2.0.1
simple-parsing==0.1.6
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
speedtest==0.0.1
speedtest-cli==2.1.3
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.2.0
statsmodels==0.14.0
streamlit==1.40.1
streamlit_folium==0.20.0
tabulate==0.9.0
tb-nightly==2.19.0a20241111
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
tensorflow-probability==0.24.0
termcolor==2.4.0
terminado==0.17.0
textblob==0.19.0
tf-nightly==2.19.0.dev20241106
tf_nightly_intel==2.19.0.dev20241106
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.1
tqdm==4.66.5
traitlets==5.5.0
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==5.2
urllib3==1.26.9
uuid==1.30
validators==0.22.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchdog==4.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==3.0.9
yarl==1.15.2
zipp==3.16.1
